Fix MySQL division to return Numeric for exact decimal arithmetic (#2014)
Changed MySQL mode division operator to return Numeric (f64) instead of
Float (f32) for proper MySQL DECIMAL type compatibility. This ensures
exact decimal arithmetic matching MySQL's behavior.

Changes:
- Updated division.rs to return Numeric in MySQL mode (lines 37, 68)
- Fixed all affected unit tests to expect Numeric instead of Float
- Updated SQL_COMPATIBILITY_MODE.md documentation to reflect implementation
- Marked High Priority recommendation as completed

All 958 executor tests passing.

Fixes #2013

+++ crates/vibesql-executor/src/tests/expression_eval.rs
@@ -252,8 +252,8 @@ fn test_eval_division() {
         right: Box::new(vibesql_ast::Expression::Literal(vibesql_types::SqlValue::Integer(4))),
     };
     let result = evaluator.eval(&expr, &row).unwrap();
-    // Division returns Float for integer operands (SQLLogicTest behavior)
-    assert_eq!(result, vibesql_types::SqlValue::Float(5.0));
+    // Division returns Numeric for integer operands in MySQL mode (exact decimal arithmetic)
+    assert_eq!(result, vibesql_types::SqlValue::Numeric(5.0));
 }
 
 #[test]
